requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
together==1.2.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
